% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{search_point}
\alias{search_point}
\title{Method of Moments Estimator for gamma dist.}
\usage{
search_point(clientid)
}
\arguments{
\item{clientid}{the given sample to calculate the estimate of the parameters.}
}
\value{
data.frame
}
\description{
search_point function will calculate mom estimate for the shape parameter "k" and the scale parameter "theta" by using given sample.
}
\examples{
clientid <- '00000220d02544fffefbc4bf'
search_point(clientid)
}
