% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{search_point}
\alias{search_point}
\title{Method of Moments Estimator for gamma dist.}
\usage{
search_point(sclist, from, ...)
}
\arguments{
\item{sclist}{: client list}

\item{from}{: start date}

\item{...}{: end date}
}
\value{
data.frame
}
\description{
search_point function will calculate mom estimate for the shape parameter "k" and the scale parameter "theta" by using given sample.
}
\examples{
sclist <- c('00000220d02544fffefbc4bf', '00000220d02544fffefc768c');
search_point(sclist, '2019-09-01', '2019-09-12')
}
